Safari-611.1.10/LayoutTests/media/modern-media-controls/macos-inline-media-controls/macos-inline-media-controls-controls-bar-styles-expected.txt

Testing the MacOSInlineMediaControls computed styles.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS style.left is "6px"
PASS style.bottom is "6px"
PASS style.width is "668px"
PASS style.height is "31px"
PASS bounds.left is 6
PASS bounds.top is 263
PASS bounds.width is 668
PASS bounds.height is 31

PASS successfullyParsed is true

TEST COMPLETE
